// kubernetes/generatedHelmChartsTsTypes/meilisearchMeilisearch.ts
// Don't Edit. This is autogenerated.
export interface IMeilisearchmeilisearch {
    replicaCount: number;
    readinessProbe: ReadinessProbe;
    livenessProbe: ReadinessProbe;
    image: Image;
    nameOverride: string;
    fullnameOverride: string;
    environment: Environment;
    auth: Auth;
    serviceAccount: ServiceAccount;
    podAnnotations: Annotations;
    customLabels: Annotations;
    service: Service;
    container: Container;
    ingress: Ingress;
    persistence: Persistence;
    resources: Annotations;
    nodeSelector: Annotations;
    tolerations: any[];
    affinity: Annotations;
    command: any[];
}
interface Persistence {
    enabled: boolean;
    accessMode: string;
    size: string;
    annotations: Annotations;
    volume: Volume;
}
interface Volume {
    name: string;
    mountPath: string;
}
interface Ingress {
    enabled: boolean;
    className: string;
    annotations: Annotations;
    path: string;
    hosts: string[];
    tls: any[];
}
interface Container {
    containerPort: number;
}
interface Service {
    type: string;
    port: number;
    annotations: Annotations;
}
interface ServiceAccount {
    create: boolean;
    name: string;
    annotations: Annotations;
}
interface Annotations {}
interface Auth {
    existingMasterKeySecret: string;
}
interface Environment {
    MEILI_NO_ANALYTICS: boolean;
    MEILI_ENV: string;
}
interface Image {
    repository: string;
    tag: string;
    pullPolicy: string;
    pullSecret?: any;
}
interface ReadinessProbe {
    periodSeconds: number;
    InitialDelaySeconds: number;
}